Support dynamic quantum architecture for IQM QPUs (#3298)
* Support dynamic quantum architecture for IQM QPUs

For IQM machines fetch the quantum architecture from server, derive the dynamic architecture from this and write it to an architecture file. This file is then passed in the lowering pipeline in place of the previously used static architecture files.

The output from the algorithm for finding usable qubits generates a list of qubits which are calibrated for mz, cz and prx. A mapping ensures that qubit names are enumerated in a linear manner and not fully calibrated qubits do not show in the architecture file. The architecture file is created in the system temp directory with a unique filename and deleted after the job has been sent to the IQM server.

Changes in function
* Removed parameter "--iqm-machine" which was used to specify the IQM quantum architecture at compile time.
* Removed static architecture files provided for Adonis, Apollo, and Aphrodite architectures.
* Removed machine architecture bindings from IQM tests.
* Implemented getting the authorization token from IQM_TOKEN environment variable. If this is not set the previous mechanism reading the token from a JSON file is used.
* Updated IQM JSON output of LLVM to latest syntax. "phased_rx" -> "prx", "measurement" -> "measure" in both emitter code and tests.

Test related changes
* Adapted IQM mock server to latest iqmclient API version (v28.0.0) and IQM server API changes.
* Implemented in IQM mock server the 20 qubit Apollo quantum architecture.
* Added support for mapping of qubits syntax in the job request to the mock server.
* In the mock server deliberately removed QB2 and QB3 from the list of PRX calibrated gates. This results in changes in the topology and forces the CUDA-Q code to work with an imperfect calibrated system. This tests the implemented algorithm with every testcase which uses more than 1 qubit.
* In mock server implemented a full checking of gates used in circuits versus the topology to verify that all 2 qubit gates are supported by the mock qpu.
* Pass the authorization token to IQM mock server with env-variable. This includes also a new testcase addressing the IQM_TOKEN environment variable.
* Save and restore environment variables for testcases modifying them. So the order of testcases does not matter anymore.
* Fixed evaluation of 2 qubit testcases which failed on some simulations.

Documentation
* Updated the IQM parts of the documentation and examples. Explained changes from previous version and removed references to the no longer needed machine names.
* Cleanup: improved doxygen headers and few other comments.
* Cleanup: Added copyright to touched files

Addresses issues: NVIDIA#1589, NVIDIA#865

---------

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>
Signed-off-by: iqm-bhoffmann <bernd.hoffmann@meetiqm.com>

* removing yapf as spell checker is complaining about it

Signed-off-by: Sachin Pisal <spisal@nvidia.com>

* Removed debug prints which are failing

Commented out debug output dumping JSON data. The code worked earlier but now fails compilation reporting a type issue.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Removed commented-out debug prints entirely

Instead of fixing a formatting error removed some debug prints entirely as they are just details of a more general previous printout.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Fixed emulation mode for IQM server

Fixed implementation of --emulate parameter. Previously it tried load the dynamic quantum architecture from server despite "emulate" means without server. This caused several ctests to fail.
Removed the --iqm-machine parameter from all ctest lines in the various C++ test programs.
Adapted the ctests mapping testcase to use a Crystal_5.txt file and restored this to the project.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Fixed mapping file issue for emulation mode

In emulation mode provide a mapping file for the 20 qubit crystal architecture.
Restored the Crystal_20.txt (former Apollo.txt) mapping file. This is used only in emulation mode.
Added header file <unistd.h> which was missing in IQMServerHelper.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Corrected comments and code readability

After feedback in internal review corrected:
- Wording of some comments.
- Improved writing of the bit encoded flag variable. Only for readability - non functional change.
- Removed an unused include.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Correction from clang-format

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Remove again `prx` from spellchecker whitelist

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Clarified documentation

Review feedback: Removed outdated paragraph.
Fixed typo in example.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Alternate ways to define the IQM QPU architecture

Allow defining the IQM QPU architecture at runtime via environment variable IQM_QPU_QA. The value of this variable is interpreted as a path+filename of an QPU architecture (mapping) file.
Aternatively the IQM QPU archtecture can be given at compile time with parameter "--mapping-file" to the nvq++ or in the backend-string as "mapping_file".
Moved the IQM QPU architecture files from "runtime" into the "targettest" folder. Added file for the IQM Crystal-54 architecture. Changed the lit variable pointing to these folders.
Updated all testcases testing IQM target which use "--emulate" to use QPU architecture via environment variable.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Updated IQM doc, save DQA to file, tests

Updated the IQM backend documentation. Created a new subpage there to explain some advanced use-cases.
Added a way to save the dynamic quantum architecture of a server to file for later use.
Added example files of a quantum architecture to targettests/Target/IQM/ folder.
Added testcase to unittests testing the options to load and save a dynamic quantum architecture to file.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

* Fixup lint and linkcheck issues

The link to Resonance is valid but only when requested with an accept string of "text/html" a response with 200 will be returned. Reverted to previously page instead.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>

---------

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>
Signed-off-by: iqm-bhoffmann <bernd.hoffmann@meetiqm.com>
Signed-off-by: Sachin Pisal <spisal@nvidia.com>
Co-authored-by: Sachin Pisal <spisal@nvidia.com>
Co-authored-by: Bettina Heim <heimb@outlook.com>